+++ src/components/CarouselSection/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from '../../styles/CarouselSection.module.css'
 
 
-export default function Carousel(){
+export default function CarouselSection(){
 
     return(
         <section id="carousel" className={styles.carouselSection}>
