+++ src/components/ContactSection/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from '../../styles/ContactSection.module.css'
 
-export default function Contact(){
+export default function ContactSection(){
 
     return(
         <section id="contact" className={styles.contactSection}>
