+++ src/components/Layout/FooterMain.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from '../../styles/FooterMain.module.css'
 
 
-export default function Footer(){
+export default function FooterMain(){
 
     return(
         <footer className={styles}>
